Extract Feedback and Logging info into separate chapter
In this commit I extracted the information about feedback and logging
out of the querying chapter and into its own chapter. I also extended
the information about providing feedback, which was pretty much
non-existent before. It now provides the prerequisites to the
experimentation chapter, which also builds on the notion of feedback.