Update to gitbook 2.6.4
This commit updates the gitbook version from 1.5 to 2.6.4. More
precisely we are specifying 2.x.x. The npm dependency is now
gitbook-cli, which provides the gitbook command. The gitbook version
itself is specified in book.json.
As part of this chaneg we also add the anchor plugin, which adds link
symbols to headings on hover that can be used to extract anchor URLs
linking to the respective heading.