Consolidate data import chapters
This change moves the chapters about RDDs and ODBC into the "Working
with Data" chapter and removes the separate "Data Formats and Sources"
chapter, which appeared to be redundant.

Note that this change only affects content of pages and the TOC; the
file structure has not been changed, so potential links to our content
wont't break.